requirements.txt
gymnasium
numpy
pyglet
miniworld
ratinabox
hydra-core
tqdm
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117